fix: adding hashcode usage changes in vectors

+++ java/memory/memory-core/src/main/java/org/apache/arrow/memory/util/ByteFunctionHelpers.java
@@ -295,7 +295,7 @@ public class ByteFunctionHelpers {
   }
 
   /** Compute hashCode with the given {@link ArrowBuf} and start/end index. */
-  public static int hash(final ArrowBuf buf, long start, long end) {
+  public static long hash(final ArrowBuf buf, long start, long end) {
 
     return hash(SimpleHasher.INSTANCE, buf, start, end);
   }
@@ -303,7 +303,7 @@ public class ByteFunctionHelpers {
   /**
    * Compute hashCode with the given {@link ArrowBufHasher}, {@link ArrowBuf} and start/end index.
    */
-  public static final int hash(ArrowBufHasher hasher, final ArrowBuf buf, long start, long end) {
+  public static final long hash(ArrowBufHasher hasher, final ArrowBuf buf, long start, long end) {
 
     if (hasher == null) {
       hasher = SimpleHasher.INSTANCE;
@@ -313,7 +313,7 @@ public class ByteFunctionHelpers {
   }
 
   /** Generate a new hashCode with the given current hashCode and new hashCode. */
-  public static int combineHash(int currentHash, int newHash) {
+  public static long combineHash(long currentHash, long newHash) {
     return currentHash * 31 + newHash;
   }
 }
